moved the last of plugins to world dir
