--- helios_sim/src/plugins/world/README.md
@@ -1,10 +1,0 @@
-# World Spawner
-
-Followed this tutorial: https://www.youtube.com/watch?v=JuVfA3V3DVU
-
-Basically, I have to create two files. One for the scene and one for the mesh
-
-Reason why is that the meshes from the regular scenes do not match well with the coordinate systems whenever bevy abstracts the meshes when creating collisions. 
-
-TODO: write more about what the tutorial describes to do.
-TODO: figure out how to add roads, water, or non-collision objects but not have collisions those textures etc.
