+++ helios_sim/src/world/atmosphere.rs
