+++ helios_sim/src/world/object_helpers.rs
