+++ helios_sim/src/world/objects.rs
